Rename dashboards and fix pictures by admin page
