+++ mysneaker/src/components/LehrerDashboard/KlasseContainer/charts/ResearchInvest.js
