+++ mysneaker/src/lib/API/src/Exceptions/NewEntryException.js
@@ -1,0 +1,5 @@
+export class NewEntryException implements Error {
+    constructor(message) {
+        this.message = message
+    }
+}
